server/backend/requirements.txt
aiosqlite==0.21.0
alembic==1.14.0
anthropic>=0.45.2
arize-phoenix-otel==0.8.0
asgiref<4.0.0,>=3.8.1
async-timeout==5.0.1
asyncpg==0.30.0
bidict==0.23.1
boto3==1.35.99
celery==5.4.0
docx==0.2.4
email-validator==2.2.0
fastapi==0.110.2 
fastapi-users==14.0.1
fastapi-users-db-sqlalchemy==7.0.0
flower==2.0.1
greenlet==3.1.1
httpx-oauth==0.16.1
langchain-anthropic==0.3.9
mixedbread-ai==2.2.6
numpy==2.1.3
odfpy==1.4.1
openinference-instrumentation==0.1.23
openinference-instrumentation-langchain==0.1.35
openinference-semantic-conventions==0.1.14
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
orjson==3.10.15
odfpy==1.4.1
packaging==24.2
pgvector==0.3.6
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pytest==8.3.5
pytest-asyncio==0.25.3
python-docx==0.8.11 
python-engineio==4.11.2
python-socketio==5.12.1
pyjwt[crypto]>=2.9.0,<2.11.0
redis>=5.0.0,<6.0.0
requests==2.32.3
scipy==1.14.1
SQLAlchemy[asyncio]==2.0.36
typing-inspect==0.9.0
typing_extensions==4.12.2
uvicorn[standard]<0.30.0,>=0.29.0
voyageai==0.3.2
websockets==12.0